Alternative ERC20 implementation with Table
This PR adds the Table module and an alternative ERC20 implementation using Table.

Most of the functions in the Table module are declared to be native. They may be compiled into the Solidity-like maps by the Move-to-EVM compiler. Moreover, although it is slow, it's possible to run them on MoveVM today because they can be easily implemented using vector in Move. It can become faster in the future once Move supports an efficient map type.

As for the ERC20 contract, the differences from the original implementation (Token.move) are as follows:

- Table is used
- All data structures go to the single struct called State (which correspond to the contract state in Solidity)
- State is published when the contract is deployed and the constructor is called.
- signers are not required for adding entries to the tables
- In order to make the names compatible, the camel case is used instead of the snake case.
- events are emitted according to the EIP20 standard.

Closes: #40