[move-to-yul] Implement Destroy instruction
This implements the destroy instruction (for structs for now only, as vectors are not implemented yet).

A few more minor fixes are included as well, which cause larger baseline changes. Specifically, some instruction implementations did not properly printed the source location, which lead to some confusing output in the baseline files.

Closes: #55