[move-to-yul] Semantic tests
This PR adds the testsuite infra to execute semantic tests. For now only two simple arithmetic tests are implemented in Arithm.move. The approach is as follows:

- A new attribute `#[evm_test]` is supported. It can be attached to a parameterless function in the Move sources.
- The compiler has a new mode to generate code for tests. If this is enabled, it will generated for each function marked with `#[evm_test]` a standalone contract (without any deployment code) which just executes this given function.
- These per-function tests will be generated in our existing `testsuite.rs` and using the `evm-exec-utils`, being executed as a custom function. The result of this execution is either success or revert, and will be dumped to the baseline.

The PR also refactors the way how we generate contracts. It turns out that solc only accepts a single Yul object as input. We now generate (in non-test mode) a single contract which contains `#[callable]`s from any modules which are compilation target. (That is, no use any longer for `#[contract]` attribute on module level. Because of this and other changes, a lot of baseline changes happened. It's best to focus on the ones in `Arithm.exp` to see the way how the semantic tests are wired up.

Closes: #77