[move-vm] Extension mechanism for VM state
This adds an extension mechanism to the Move VM which allows native functions to access some state define by the execution environment. The mechanism is based on the Rust `Any` type. New variantes of the execute functions in the session (`execute_XXX_with_env_ext`) which allow to pass in the extension which can then be consumed by natives.

This PR has been pulled out of #142 for better transparency of this new mechanism. No tests are in this PR as the VM doesn't have its own test suite but is tested indirectly via other creates; #142 contains a number of tests
for this.

Closes: #144