[move-lang][vm] Allow non-`'static` native context extensions. (#3)
[*Note*: in addition to the described changes, this PR fixes various issues for CI in the new repo (to the extend they surfaced here). A new `profile.ci` was introduced which allows to fine tune build in CI to pass the current restrictions there regards storage. This required upgrading the tool chain and fixing follow-up issues.]

This uses the `Tid<'a>` type from the [better_any](https://docs.rs/better_any/latest/better_any/) crate which acts as a replacement for the standard `Any` type in order to implement `NativeContextExtensions`. This allows the polymorphic values stored in the extensions to be non-`'static`, which `Any` does not.

As a consequence, it is now `NativeContextExtensions<'a>`, where `'a` is placeholder for the lifetime of any references stored in some extensions. Only one lifetime parameter is allowed for such extensions because of restrictions of `better_any`.

A user of extensions should not see any of this, however, structures describing extensions must use `#[derive(Tid)]`, and any type parameters those have must do as well. This is so the crate can compute a `TypeId` for such types.

The `NativeContextExtensions<'a>` type has been moved in its own module, and a test has been added to verify that this new feature works.