[ci] Run CI on merge commit instead of head commit
Running on head commit increases the chance that CI run passes standalone but fails once the changes are merged.

Also removes checking out all branches because that seems to be not needed in the Move repo.

Moreover, this adds a new `./scripts/check_pr.sh` to check whether a PR is ready for CI.