[file-format] Introduce metadata in file format (unused for now) (#123)
* [file-format] Introduce metadata in file format (unused for now)

This PR adds a generic representation of metadata into the file format.

We want to do this as soon as possible because it is a breaking change, and before more Move code hits mainnet.

The design is rather simple: metadata is represented as a vector of pairs of a key and a value, both byte blobs. A convention to organize keys is to be defined outside of the file format, one approach is to make them 32 byte hashes over a naming scheme. This allows us to store attributes (in serialized form) as well as other metadata we need over time. For comparison, in EVM bytecode there is only a single byte blob for metadata. The keys we have allow to search the metadata for a specific value without the need of decoding it all.

* Adding an `--bytecode-version <n>` option to the compilation chain.

The option can be provided to the compiler command line or fed into the package system via the env var `MOVE_BYTECODE_VERSION`. Valid values are 5 (file format before metadata) and 6 (current),

* [fixup] Addressing reviewer comments

* [fixup] Merging changes into V5 instead of a new V6