[scripts] Added Git installation to dev_setup.sh
Although most of the development setups have Git installed (since
development happens in a Git repository), this assumption does not
hold for all the cases: for example in the case of Docker's "ADD ."
where local files are copied into a container without Git.

However, Git is required by the Move CLI, although this is
undocumented.

Hence, the development setup script should have an option for
installing it (although it's disabled by default).