[stdlib] UTF8 String type
This introduces a new type `string::String` which represents a valid UTF8 string. The type is intended to replace `ascii::string`.

The UTF8 enforcement is based on Rust's and correct only modulo Rust. When creating a string with `string::utf8(bytes)`
the Rust `str` implementation will check the correctness of the encoding. On deconstructing or composing strings, the Rust function
`str::is_valid_char_boundary` is used to allow composition only on utf8 boundaries. The String type does not allow direct mutation of
bytes, therefore utf8 correctness is not violated in other ways.

Currently, the only way to denote a string is as in `string::utf8(b"hello")`. If someone would want to create unicode characters, they
would have to use the `\x1FH` escape in the Move string syntax. We may want to consider introducing string literal `u"foo"` which can directly be used to
denote strings, and which also supports `\u{digits}` style of unicode escapes.