[unit-test] Allow extensions used for unit testing to live outside of Move repo
Until now, extensions leveraged by the Move UT infra had to live in the Move repo.

This PR adds a hook which can be used as in `move_unit_tests::extensions::set_extension_hook(fun)`,
where `fun` is populating a `NativeContextExtensions` for the according custom environment. How
to use this is documented in the code at `extensions.rs` and also in an integrated unit test.