[Move analyzer] Updated package info & clap usage
While "move-analyzer --version" reports "1.0.0", the package itself
is at version 0.0.0. This is because version information is hardcoded
to the clap() usage, and the Cargo.toml package manifest is
out-of-date.

Updated Cargo.toml information: dropping the Diem repo and setting a
proper version. Clap() usage was also changed to idiomatic one as per
9225b84aa ("Idiomatic "clap" usage (implements --version switch for
tools) (#81)", 2022-05-02)