[move-prover] write-back following the whole ancestor tree
The current memory model in Move Prover has a flaw in writing back
dying references, as identified in issue #299.

The gist of the flaw is this:
we can only write-back a reference on code paths where it is guaranteed
to be defined. Otherwise in the prover's point of view, the write-back
may cause arbitrary value to be written back to the parent `Mutation`
value and cause unexpected verification failures.

As a side note, this is also visible in the concrete executor of the
spec + move (i.e., the stackless VM).

The solution, as implemented in this commit, is to ravamp the way
write-backs are done.

In summary, the solution is to write-back the dying references according
to DFS paths instead of BFS layers.

Taking the example in #299:

The borrow graph looks like this:

```
     --borrows_from--> t7 --borrows_from--> Global(t1)
t3 --|
     --borrows_from--> t8 --borrows_from--> Global(t2)
```

And the old approach is to write-back according to BFS layers,
i.e., t3, then t7 and t8, which yields the following psudo-code.

```
if IsParent[t7](t3) {
    WriteBack[t7](t3);
}

if IsParent[t8](t3) {
    WriteBack[t8](t3);
}

// both t7 and t8 are not always defined when code reaches here
WriteBack[Global](t7);
WriteBack[Global](t8);
```

The correct behavior should be DFS paths, i.e.:

```
if IsParent[t7](t3) {
    WriteBack[t7](t3);
    // t7 is guaranteed to be defined when code reaches here
    WriteBack[Global](t7);
}

if IsParent[t8](t3) {
    WriteBack[t8](t3);
    // similarly, t8 is guaranteed to be defined in this block
    WriteBack[Global](t8);
}
```