[move-prover] fix is_parent tests for multiple write-back chains
PR #302 has an error in terms of placing `IsParent` checks for different
borrow chains.

Consider the following case:

```
                        --borrows_from--> C
A --borrows_from--> B --|
                        --borrows_from--> D
```

When references A, B, C, D go out of scope, this will yield two chain of
references:

A --> B --> C and
A --> B --> D

This is correctly captured in #302.

The bug is how to generate the `IsParent` test:

The correct behavior is the following:

```
if IsParent[C](B)
    WriteBack[B](A)
    WriteBack[C](B)

if IsParent[D](B)
    WriteBack[B](A)
    WriteBack[D](B)
```

i.e., branching at node B.

But in #302, I mistakenly used the root of the chain as the branching
point and hence, both the tests become `IsParent[B](A)`, which is not
correct.

This commit fixes this issue by deriving the deterministic factor for
the write-back chain.

Consider the case with the borrowing graph:

```
                  ---> E
          ---> C -|
A ---> B -|       ---> F
          ---> D
```

This will yield three chains:
```
if IsParent[C](B) && IsParent[E](C)
    WriteBack[B](A)
    WriteBack[C](B)
    WriteBack[E](C)

if IsParent[C](B) && IsParent[F](C)
    WriteBack[B](A)
    WriteBack[C](B)
    WriteBack[F](C)

if IsParent[D](B)
    WriteBack[B](A)
    WriteBack[D](B)
```