[move-prover] use a special location to mark uninitialized mut refs
Mutable references cannot be arbitrarily initialized, they must be
initialized into a state where the IsParent test have to fail if this
reference is not assigned otherwise.

To be specific, consider the folowing example:

```
let v1 = 0;
let v2 = 0;

let r = if (cond) {
    let r1 = &mut v1;
    r1
} else {
    let r2 = &mut v2;
    r2
};
```

This will be translated to:

```
r1 := $uninitialized();
r2 := $uninitialized();

... <omitted> ...

if (cond) {
    r1 := borrow_local(v1);
    r := r1
} else {
    r2 := borrow_local(v2);
    r := r2
}

... <omitted> ...

if (IsParent[r1](r)) {
    WriteBack[r1](r);
    WriteBack[v1](r1);
}
if (IsParent[r2](r)) {
    WriteBack[r2](r);
    WriteBack[v2](r2);
}
```

It is very important to mark `r1` and `r2` and `$uninitialized()`
to ensure that there will be one and only one `IsParent` test activated
when the write-back happens.

The previous approach is to `assume IsEmptyVec(p#$Mutation($t_i));`
for all `t_i` that is a mutable reference used in function locals.

However, setting the `path` vector in the `Mutation` struct to zero
does not guarantee that this mutation, when put under `IsParent` test,
is not going to return false when that mutation is not initialized.

As can be seen in the example above, even after proper assignment, `r1`
and `r2` both have a `path` equal to the length of zero... (because `r`
is borrowing from `r1` and `r2` `as direct borrow, not field or element
borrow).

This commit fixes this issue with an explicit `$Uninitialize()`
constructor for the `$Location` part of a `$Mutation`.