[move-prover] do not peephole optimize the initial local and global borrows
Live variable analysis allows the following peephole optimization:

```
  $t := call(...)
  x := $t
  <$t is dead>

  ~~~

  x := call(...)
```

This optimization makes perfect sense but unfortunately conflicts with
one policy later: IsParent test can only be applied on two references:

Taking the following code as an example:
```
let v1 = 0;
let v2 = 0;

let r = if (cond) {
  let r1 = &mut v1;
  r1
} else {
  let r2 = &mut v2;
  r2
};
```

References `r1` and `r2` will be optimized away after the live var
analysis. This leaves the reference `r` pointing to two local locations.

And later when translating `IsParent` test to Boogie, we will run into
cases like:

```
IsParent[LocalRoot(v1)@](r)
IsParent[LocalRoot(v2)@](r)
```

And this violates our policy that IsParent can only be applied among
two references.

As a result, we disable this optimization for the initial local/global
borrowing, which leaves the IsParent tests to be
```
IsParent[Reference(r1)@](r)
IsParent[Reference(r2]@](r)
```

An alternative solution to this problem is to implement IsParent tests
between a reference to a `LocalRoot` and a `GloablRoot`. But that is a
more significant change and does not buy much benefits performance-wise.