[move-prover] use an `Uninit` operation to mark uninitialized mut refs
Following from the above commit, instead of marking all mutable
references in the function locals (except the function arguments) as
uninitialized, this commit further restrict the number of mutable
references that can be marked as `Uninit`, hence, reducing the chances
of introducing inconsistent assumptions in the verification.

The basic idea is to leverage the livevar analysis results, and mark
whatever mutable reference that is
1) alive even before the first instruction of the function and
2) is not a parameter,
as uninitialized.

This commit also includes a fix to the concrete spec executor such that
we do not run into errors when a mutable reference is mysterously
destroyed even before it is initialized.