[move-prover] do not let bind a mutable reference for `EmitEvent`
```
public fun emit_event<T>(handle_ref: &mut EventHandle<T>, msg: T) { ... }
```

Events are verified in Move Prover in a special manner. One example is
that the `EmitEvent` bytecode is especially designed to simulate the
`emit_event` function.

During the invocation of `EmitEvent`, we might be let-binding a mutable
reference which we really do not want as introcuing a mutable reference
via `Assume(Identical)` complicates the live var and borrow analysis
(for example, we will treat the newly bounded reference as uninitialized).

While essentially we do not need a reference to the handle and we can
just let-bound the `$Dereference ( <exp> )` value directly, which is
what is essentially done in this commit.