[move-cli] Added "docgen" command (#291)
* [move-cli] Added "docgen" command

Move Prover's "--docgen" switch is not currently exposed to the user,
so added a top level command "docgen" to do the trick.

Also forwarding --docgen-template if needed, and implemented a new
switch --quiet mainly for test setups where output must be
deterministic. The --docgen-template switch is currently the only
Docgen related option switch exposed to the user by the Prover.

Move Prover is also mentioned in the description, so the user knows
they must have a working Move Prover setup in order to use this
feature.

* [move-cli] Added a test for "docgen" command

Added a simple test for the new "move docgen" command.

We are testing a simple document generation with a simple template,
supressing the output for testing purposes (with the --quiet switch).