[move-prover] remove the destroy instruction for imm borrow removal
The `destroy($t1)` bytecode is not removed in the immutable borrow
pass, which it should be, otherwise we will observe an `destroy($t1)`
where `$t1` is not a reference type.

This does not cause any issue with the symbolic reasoning as `destroy`
is essentially translated into a no-op. Instead, it is caught by the
concrete executor.