[vm] Invalidate and auto-flush code loader cache
This is a hack to workaround well-known issues with the VMs code loader cache. This cache is not designed to support module upgrade or deletion. This leads to situations where the cache does not reflect the state of storage:

1. On module upgrade, the upgraded module is in storage, but the old one still in the cache.
2. On an abandoned code publishing transaction, the cache may contain a module which was never committed to storage by the adapter.

The solution is to add a flag to `Loader` marking it as 'invalidated'. For scenario (1), the VM sets the flag itself. For scenario (2), a public API allows the adapter to set the flag.

If the cache invalidated, it can (and must) still be used until the end of the session. This is because during session execution, data might have been extracted which indirectly refers to the current cache (e.g. `CachedStructIndex`). Rather, when a *new* session is created, the VM will check the state of loader and flush it if needed. Note that this solution implies no concurrent excution of sessions.