[move-prover] do not remap moved nodes in borrow analysis (#335)
* [move-prover] do not remap moved nodes in borrow analysis

If a local is first mutably borrowed and later moved in a loop, the
reference will have two parents (and hence, write-back edges):
- one to the original local and
- the other to the LHS of the moved local.

This can be shown in the following code:

```move
loop {
  let s = S { x : 100  };
  let r = &mut s;
  if (r.x >= limit) {
    r.x = limit;
  }
  t = move(s);
}
```

The borrow analysis will conclude that the reference `r` has two
parents, `s` and `t`.

This bug can only happen in loops due to the fact that the borrow
analysis is essentially an abstract interpretation pass and the
join operator for the `BorrowInfo` is a set union. Although for any
single pass on the loop body, `r` can either point to `s` or `t`. After
the join, `r` points to both `s` and `t`, which is incorrect.

This commit fixes this issue by dropping the `remap` function on
borrowed edges for a moved node. Specifically:

```
t = move(s);
```

- Prior to this fix, all references derived from `s` will be remapped
to be deriving from `t` (which does not seem to be correct).
- After this fix, the references derived from `s` remain pointing to `s`

During the process, the `moved_nodes` in the `BorrowInfo` seems to be
obsolete as well, so it is removed, which also causes the `exp` test
cases to be updated.`

* fixup! [move-prover] do not remap moved nodes in borrow analysis