[packages] New dependency type 'node' + custom package info properties
This adds a new dependency type to the package manifest which allows to specify a node from which to receive a package:

```
[dependencies]
MoveStdlib = { node: "localhost:8080", address: "0x1", package: "move-stdlib" }
```

Resolution of such kind of a dependency is handled via a new trait `PackageHooks` which can be globally registered on a per-process basis. If no hook is registered, an error will be produced that `node` deps are not supported.

Note: After investigating some time whether the resolver could be passed in as a parameter, it became clear that a larger refactoring woul be needed as the package system is simply not designed for compositionality like this.

In additionn to the new dependency type, this also allows additional fields in the `[package]` section, as in

```
[package]
name = "Foobar"
upgrade_policy = ..
```

The valid extra names are again determined by the `PackageHooks`.