[move-prover] switch src/dst pair for the havoc operator
The `havoc` operator has been modeled as:
```
() := havoc($i);
```

i.e., 0 output and 1 input.

This commit switches the src and dest pair into
```
$i := havoc();
```

This is in fact closer to the semantic of `havoc`: it re-assigns the
target with an arbitrary value. therefore, the target should be on the
receiving end.

This has important effect on our program analysis passes, especially
reaching definition analysis and live-var analysis.