[compatibility] Relax compatibility for friends
This adds a dynamic flag which, if set, lets the compatibility check treat friend functions like private functions. The feature
is not on by default because it assumes that modules are compiled and published on package granularity level. On package level, we can assume
that each call site of the friend function has been updated to the changed friend function as well.