[move-model] do not report errors when translating a non-pure fun to spec fun
During the process of translation, if an error already happened, i.e.,
the move function is already considered non-pure, the following-up
errors can be misleading. As shown in the test case.

This commit stops showing an error on unable to find the correct spec
fun candidate due to type unification failures (which is already messed
up when there is a previous error).