[compiler] Resolve 0x1::vector functions to opcodes
The phase in the compiler which maps calls to vector functions to opcodes only catches cases like `std::vector::fun`. If a user has `0x1::vector::fun` (which he should be free to do) no resolution to opcodes happens, and the legacy native functions in the library are called. Those functions, however, should be deactivated as they represent a further surface of concern of security and gas metering.

This PR lets the vector functions resolve to opcodes if either (a) the symbolic address name is `std` (b) or the address resolves to `0x1`.