[fuzzing] Add fuzzing infra for bytecode verifier + fix bug (#564)
This adds a fuzzing crate and a few targets for `CompileModule` to the bytecode verifier.
The rather rudimentary `cargo-fuzz` is used ([see here](https://rust-fuzz.github.io/book))
as the more sophisticated tools like `honggfuzz` don't work on Apple silicon. Using this
(and other fuzzing tools) requires to derive `arbitray::Arbitray` (not to be mistaken
with proptests `Arbitrary`) which allows to derive values from bit images, as the fuzzer
produces.

This fuzzer infra was capable to find a recent (already fixed) bug in stack balancing in reference analysis,
caused by continuing analysis after an error occured. It also discovered (for now) the following
new bug: if the number of locals + parameters is larger than 256 certain analysis will go wrong,
as `type LocalIndex = u8`. This was discovered since reference analysis uses `0..self.locals as LocalIndex`
to iterate over locals, which reduces to `0..0`. It appeared that bounds check actually does
not check whether the number of locals + parameters in range, which was added in this PR.