[binary/cfg] Relax backward-jump == loop back-edge requirement
When generating a `VMControlFlowGraph`, identify loop heads and
back-edges by finding back-edges in the depth-first traversal from the
entry block.

This is roughly the algorithm proposed for finding widening points in
Section 4.1 (Depth-first numbering) of "Efficient chaotic iteration
strategies with widenings", Bourdoncle 1993.

Replaces the previous strategy of assuming a backward (in code
offset) jump corresponds to a back-edge, which is something the
Control Flow verifier was ensuring.  This assumption is no longer
required, so this commit also removes references to that dependency.

Test Plan:

Newly introduced unit tests check that a CFG can be generated, even
when the verifier would fail on the bytecode:

```
move/language/move-binary-format$ cargo test
```