[rfc][bytecode verifier] simplifications enabled by 'return on first error' discipline (#561)
92d1d29efa9ec458938599c6524a28287e44e2e5 fixed a bug in the reference safety verifier pass by returning the first error encountered instead of collecting all errors, then picking one. This change shows that the verifier is basically trying to implement a Rust `Result` in a very indirect way: aggregate a bunch of errors into an `InvariantMap`, let the client iterate through them and pick one, then turn it into a Rust `Result`.

This PR changes the abstract interpreter to use a Rust `Result` directly via the following changes:
- Return a `Result` instead of an `InvariantMap` from `analyze_function`. All call sites were iterating through the `InvariantMap` to see if there was an error, but the caller will now get the error directly.
- Get rid of `BlockPostcondition`. This is no longer needed now that we return the error directly
- Get rid of the `AnalyisError` associated type and use `PartialVMResult` directly. This abstract interpreter is only intended for usage inside the bytecode verifier.

Any changes to the verifier are very sensitive, so this should be looked at very carefully. These changes do not need to go in quickly because they are not fixing any active problems in the verifier or elsewhere. They are made in the interest of making a key part of the verifier smaller/simpler for the long term.

Closes https://github.com/move-language/move/issues/557