[verify/cfg] Use LoopSummary/LoopPartition in control_flow verifier
Replace the checks for invalid break, continue, loop split and depth
overflow with a test for CFG reducibility based on Tarjan's
Reducibility Algorithm [0].

This test preserves the properties that we care about for efficient
abstract interpretation, which can be equivalently described as:

1. The CFG is reducible.
2. The CFG has a unique depth-first spanning tree.
3. The body of every loop in the CFG is dominated by its head.

But it is more lenient than the previous check in that it allows
backwards jumps in bytecode that aren't continues and forward jumps
(breaks) that can escape multiple loops.  As such, it also eliminates
the `INVALID_BREAK` and `INVALID_CONTINUE` status codes, and only
reports `INVALID_LOOP_SPLIT` and `LOOP_MAX_DEPTH_REACHED`.

Closes #496

[0]: Testing flow graph reducibility, R Tarjan, 1974

Test Plan:

New and old control_flow unit tests:

```
bytecode-verifier-tests$ cargo test control_flow_tests
```

New regression test for the code example described in the issue:

```
transactional-tests$ cargo test control_flow/regression_test_496
```

The regression tests and the new tests demonstrating patterns that are
now okay are confirmed to have failed before this PR.